+++ wokwi/Timer.h
@@ -7,22 +7,22 @@
 class Timer
 {
 public:
-    explicit Timer(uint32_t durationSeconds);
+    explicit Timer(uint64_t durationSeconds);
 
-    void Start(uint32_t currentTimeMs);
-    uint32_t Update(uint32_t currentTimeMs);
-    void Stop(uint32_t currentTimeMs);
+    void Start(uint64_t currentTimeMs);
+    uint64_t Update(uint64_t currentTimeMs);
+    void Stop(uint64_t currentTimeMs);
     bool IsRunning() const;
     bool IsFinished() const;
 
 private:
-    uint32_t durationMs;
-    uint32_t accumulatedMs;
-    uint32_t totalElapsedMs;
-    uint32_t startTime;
+    uint64_t durationMs;
+    uint64_t accumulatedMs;
+    uint64_t totalElapsedMs;
+    uint64_t startTime;
     bool isRunning;
 };
 
-String formatTime(uint32_t ms);
+String formatTime(uint64_t ms);
 
 #endif // TIMER_H
